requirements.txt
pytz
boto3==1.9.233
PyYAML==5.1.2